cmake支持
还差汇编支持

+++ include/type.h
@@ -1,7 +1,7 @@
 #define NULL (void*)0
 #define false 0
 #define true 1
-typedef unsigned long long size_t;
+typedef unsigned int size_t;
 typedef unsigned int u32;
 typedef unsigned short u16;
 typedef unsigned char u8;
